sprintfix:1.测试部分接口，并做修改 2.补充测试和正式环境配置文件下的RABBITMQ的相关配置 3.删除default中关于本地数据库的配置

+++ apps/good_apply/views.py
@@ -134,6 +134,7 @@ class ApplyViewSet(viewsets.ModelViewSet):
         username = request.user.username
         apply_list = req.get('apply_list')
         org_id = req.get('org_id', None)
+        valid_user_in_the_org(org_id, username)
         flag = if_secretary(request, org_id)
         apply_status = 1
         if flag:
@@ -182,7 +183,9 @@ class ApplyViewSet(viewsets.ModelViewSet):
         获取需要审核的物资列表
         """
         req_data = request.GET
+        username = request.user.username
         org_id = req_data.get('org_id')
+        valid_user_in_the_org(org_id, username)
         flag = if_secretary(request, org_id)
         if not flag:
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
@@ -309,6 +312,7 @@ class ApplyViewSet(viewsets.ModelViewSet):
         body = request.data
         org_id = body.get('org_id', None)
         username = request.user.username
+        valid_user_in_the_org(org_id, username)
         flag = if_secretary(request, org_id)
         if not flag:
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
@@ -462,6 +466,7 @@ class OrganizationMemberViewSet(viewsets.ModelViewSet):
         req_data = request.GET
         org_id = req_data.get('org_id')
         username = request.user.username
+        valid_user_in_the_org(org_id, username)
         flag = if_secretary(request, org_id)
         if not flag:
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
@@ -528,16 +533,23 @@ class OrganizationMemberViewSet(viewsets.ModelViewSet):
     @action(methods=['get'], detail=False)
     def check_user_if_groupmber(self, request):
         """检查用户是否在组成员表内"""
-        username = request.username
+        username = request.user.username
         if not OrganizationMember.objects.filter(username=username).exists():
             raise BusinessException(StatusEnums.USER_NOMEMBER_ERROR)
         return get_result({"message": "用户已有存在组"})
 
-    @action(methods=['get'], detail=False)
+    @action(methods=['post'], detail=False)
     def get_all_groupmber(self, request):
         """获取用户表所有成员"""
-        username = request.username
-        queryset = OrganizationMember.objects.filter(username=username)
+        req = request.data
+        org_id = req.get('org_id')
+        username = request.user.username
+        valid_user_in_the_org(org_id, username)
+        flag = if_secretary(request, org_id)
+        if not flag:
+            raise BusinessException(StatusEnums.AUTHORITY_ERROR)
+        # 管理员获取除了当前组内其他的用户表中的成员
+        queryset = OrganizationMember.objects.exclude(org_id=org_id)
         groupmber_list = [groupmber.to_json() for groupmber in queryset]
         return JsonResponse(success_code(groupmber_list))
 
@@ -550,6 +562,7 @@ class OrganizationMemberViewSet(viewsets.ModelViewSet):
         if org_id in settings.INNER_LIST:
             raise BusinessException(StatusEnums.INNER_CURD_ERROR)
         username = request.user.username
+        valid_user_in_the_org(org_id, username)
         flag = if_secretary(request, org_id)
         if not flag:
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
@@ -673,18 +686,18 @@ class ApplyToOrgViewSet(viewsets.ModelViewSet):
                                                                                , result=review_result_ch))
             return JsonResponse(success_code('审批成功'))
 
-        @action(methods=['get'], detail=False)
-        def approval_user_groupmber(self, request):
-            """申请加入组接口"""
-            req = request.data
-            username = request.username
-            org_id = req.get("org_id")
-            valid_user_in_the_org(org_id, request.user.username)
-            if not Secretary.objects.filter(username=username).exists():
-                raise BusinessException(StatusEnums.AUTHORITY_ERROR)
-            queryset = ApplyToOrg.objects.filter(org_id=org_id)
-            ApplyToOrg_list = [applyuser.to_json() for applyuser in queryset]
-            return JsonResponse(success_code(ApplyToOrg_list))
+    @action(methods=['post'], detail=False)
+    def approval_user_groupmber(self, request):
+        """申请加入组接口"""
+        req = request.data
+        username = request.user.username
+        org_id = req.get("org_id")
+        if not OrganizationMember.objects.filter(username=username, org_id=org_id).exists():
+            if not ApplyToOrg.objects.filter(create_user=username, apply_group_id=org_id).exists():
+                ApplyToOrg.objects.create(create_user=username, apply_group_id=org_id, status=1)
+                return get_result({"message": "已成功申请加入该组！"})
+            return get_result({"message": "已申请，无须重复申请！"})
+        return get_result({"message": "你已经在该组当中，无需再次申请"})
 
 
 class SecretaryViewSet(viewsets.ModelViewSet):
@@ -798,34 +811,42 @@ class OrganizationViewSet(viewsets.ModelViewSet):
     @action(methods=['post'], detail=False)
     def create_group(self, request):
         """创建"""
-        username = request.username
+        username = request.user.username
         req = request.data
         group_name = req.get("group_name")
         if not Organization.objects.filter(group_name=group_name).exists():
-            org = Organization.objects.create(group_name=group_name)
-            Secretary.objects.create(username=username, org_id=org.id)
-            OrganizationMember.objects(username=username, org_id=org.id)
+            Organization.objects.create(group_name=group_name)
+            org_id = Organization.objects.filter(group_name=group_name).first().id
+            Secretary.objects.create(username=username, org_id=org_id)
+            OrganizationMember.objects.create(username=username, org_id=org_id)
             return get_result({"message": "创建组成功"})
-        raise BusinessException(StatusEnums.GROUP_ISEXISTS_ERROR)
+        raise BusinessException(StatusEnums. GROUP_ISEXISTS_ERROR)
 
     @action(methods=['get'], detail=False)
     def get_all_userin_groupid(self, request):
         """获取用户所有所在组接口"""
-        username = request.username
-        """根据用户名获取他所在的组成员表查询出来"""
-        group_user = OrganizationMember.objects.filter(username=username)
-        """循环得到"""
-        group_user_id_list = [gro.to_json() for gro in group_user]
-        return JsonResponse(success_code(group_user_id_list))
+        username = request.user.username
+        # 根据用户名获取他所在的组成员表查询出来
+        group_user_list = OrganizationMember.objects.filter(username=username)
+        group_list = []
+        for group_user in group_user_list:
+            org_id = OrganizationMember.objects.filter(username=group_user.username, id=group_user.id).first().org_id
+            group = Organization.objects.filter(id=org_id).first()
+            group_list.append(group.to_json())
+        return JsonResponse(success_code(group_list))
 
     @action(methods=['post'], detail=False)
     def quit_group(self, request):
         """用户退出组接口"""
         req = request.data
-        username = request.username
+        username = request.user.username
         org_id = req.get("org_id")
         valid_user_in_the_org(org_id, request.user.username)
+        if Secretary.objects.filter(username=username):
+            return get_result({"message": "管理员无法退组，请转交管理员身份再执行相关操作"})
         OrganizationMember.objects.filter(username=username, org_id=org_id).delete()
+        Apply.objects.filter(apply_user=username, org_id=org_id).delete()
+        # GoodApply.objects.filter(username=username, org_id=org_id).delete()
         return get_result({"message": "退出组成功"})
 
     @action(methods=['post'], detail=False)
